backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
redis==5.2.1
celery==5.4.0
prometheus-client==0.21.1
python-json-logger==3.2.1
httpx==0.28.1
ruff==0.8.4
pytest==8.3.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
boto3==1.35.90
botocore==1.35.90
openai==1.59.6
oci==2.141.1
anthropic==0.42.0
